Real time prediction script added
This addition allows one to load the pretrained THOSnet model weights and do realtime classification. Be sure to position the camera upwards from a downward facing viewpoint. Also, the camera has to be a certain distance away from your hands. The model might mispredict the gestures if your hands are too close to the camera.